Menu Pan Bug
When you click on "Options" the menu should pan down room_height/2. This would move the view down 1 screen distance.
Problem: The menu ball keeps going up and down even though I set it to an alarm that shouldnt activate until it triggered by the Var_CameraActive is = 1 which only happens if options is clicked. However it seems that it just loops going up and down.